move the CSLC-specific function from `bursts` to `_cslc`
