+++ src/opera_utils/_cslc.py
@@ -1,15 +1,22 @@
 from __future__ import annotations
 
+import json
+import logging
 import re
+import subprocess
+import tempfile
 from datetime import datetime, timedelta
+from os import fspath
 from pathlib import Path
-from typing import Any, Callable
+from typing import Any, Callable, Sequence, Union
 
 import h5py
 import numpy as np
 from pyproj import CRS, Transformer
+from shapely import geometry, ops, wkt
 
 from ._types import Filename
+from .constants import OPERA_IDENTIFICATION
 
 __all__ = [
     "CslcParseError",
@@ -20,7 +27,11 @@ __all__ = [
     "get_xy_coords",
     "get_lonlat_grid",
     "get_cslc_orbit",
+    "get_cslc_polygon",
+    "get_union_polygon",
+    "make_nodata_mask",
 ]
+logger = logging.getLogger(__name__)
 
 
 CSLC_S1_FILE_REGEX = (
@@ -264,3 +275,158 @@ def get_lonlat_grid(
     lon = lon.reshape(X.shape)
     lat = lat.reshape(Y.shape)
     return lon, lat
+
+
+def get_cslc_polygon(
+    opera_file: Filename, buffer_degrees: float = 0.0
+) -> Union[geometry.Polygon, None]:
+    """Get the union of the bounding polygons of the given files.
+
+    Parameters
+    ----------
+    opera_file : list[Filename]
+        list of COMPASS SLC filenames.
+    buffer_degrees : float, optional
+        Buffer the polygons by this many degrees, by default 0.0
+    """
+    dset_name = f"{OPERA_IDENTIFICATION}/bounding_polygon"
+    with h5py.File(opera_file) as hf:
+        if dset_name not in hf:
+            logger.debug(f"Could not find {dset_name} in {opera_file}")
+            return None
+        wkt_str = hf[dset_name][()].decode("utf-8")
+    return wkt.loads(wkt_str).buffer(buffer_degrees)
+
+
+def get_union_polygon(
+    opera_file_list: Sequence[Filename], buffer_degrees: float = 0.0
+) -> geometry.Polygon:
+    """Get the union of the bounding polygons of the given files.
+
+    Parameters
+    ----------
+    opera_file_list : list[Filename]
+        list of COMPASS SLC filenames.
+    buffer_degrees : float, optional
+        Buffer the polygons by this many degrees, by default 0.0
+    """
+    polygons = [get_cslc_polygon(f, buffer_degrees) for f in opera_file_list]
+    polygons = [p for p in polygons if p is not None]
+
+    if len(polygons) == 0:
+        raise ValueError("No polygons found in the given file list.")
+    # Union all the polygons
+    return ops.unary_union(polygons)
+
+
+def make_nodata_mask(
+    opera_file_list: Sequence[Filename],
+    out_file: Filename,
+    buffer_pixels: int = 0,
+    overwrite: bool = False,
+):
+    """Make a boolean raster mask from the union of nodata polygons using GDAL.
+
+    Parameters
+    ----------
+    opera_file_list : list[Filename]
+        list of COMPASS SLC filenames.
+    out_file : Filename
+        Output filename.
+    buffer_pixels : int, optional
+        Number of pixels to buffer the union polygon by, by default 0.
+        Note that buffering will *decrease* the numba of pixels marked as nodata.
+        This is to be more conservative to not mask possible valid pixels.
+    overwrite : bool, optional
+        Overwrite the output file if it already exists, by default False
+    """
+    from osgeo import gdal
+
+    gdal.UseExceptions()
+
+    if Path(out_file).exists():
+        if not overwrite:
+            logger.debug(f"Skipping {out_file} since it already exists.")
+            return
+        else:
+            logger.info(f"Overwriting {out_file} since overwrite=True.")
+            Path(out_file).unlink()
+
+    # Check these are the right format to get nodata polygons
+    try:
+        dataset_name = parse_filename(opera_file_list[-1])
+    except CslcParseError:
+        raise ValueError(f"{opera_file_list[-1]} is not a CSLC file")
+
+    try:
+        dataset_name = parse_filename(opera_file_list[-1])
+        test_f = f"NETCDF:{opera_file_list[-1]}:{dataset_name}"
+        # convert pixels to degrees lat/lon
+        gt = _get_raster_gt(test_f)
+        # TODO: more robust way to get the pixel size... this is a hack
+        # maybe just use pyproj to warp lat/lon to meters and back?
+        dx_meters = gt[1]
+        dx_degrees = dx_meters / 111000
+        buffer_degrees = buffer_pixels * dx_degrees
+    except RuntimeError:
+        raise ValueError(f"Unable to open {test_f}")
+
+    # Get the union of all the polygons and convert to a temp geojson
+    union_poly = get_union_polygon(opera_file_list, buffer_degrees=buffer_degrees)
+    # convert shapely polygon to geojson
+
+    # Make a dummy raster from the first file with all 0s
+    # This will get filled in with the polygon rasterization
+    cmd = (
+        f"gdal_calc.py --quiet --outfile {out_file} --type Byte  -A"
+        f" NETCDF:{opera_file_list[-1]}:{dataset_name} --calc 'numpy.nan_to_num(A)"
+        " * 0' --creation-option COMPRESS=LZW --creation-option TILED=YES"
+        " --creation-option BLOCKXSIZE=256 --creation-option BLOCKYSIZE=256"
+    )
+    logger.info(cmd)
+    subprocess.check_call(cmd, shell=True)
+
+    with tempfile.TemporaryDirectory() as tmpdir:
+        temp_vector_file = Path(tmpdir) / "temp.geojson"
+        with open(temp_vector_file, "w") as f:
+            f.write(
+                json.dumps(
+                    {
+                        "geometry": geometry.mapping(union_poly),
+                        "properties": {"id": 1},
+                    }
+                )
+            )
+
+        # Open the input vector file
+        src_ds = gdal.OpenEx(temp_vector_file, gdal.OF_VECTOR)
+        dst_ds = gdal.Open(out_file, gdal.GA_Update)
+
+        # Now burn in the union of all polygons
+        gdal.Rasterize(
+            dst_ds,
+            src_ds,
+            burnValues=[1],
+            outputType=gdal.GDT_Byte,
+            creationOptions=["COMPRESS=LZW"],
+        )
+
+
+def _get_raster_gt(filename: Filename) -> list[float]:
+    """Get the geotransform from a file.
+
+    Parameters
+    ----------
+    filename : Filename
+        Path to the file to load.
+
+    Returns
+    -------
+    List[float]
+        6 floats representing a GDAL Geotransform.
+    """
+    from osgeo import gdal
+
+    ds = gdal.Open(fspath(filename))
+    gt = ds.GetGeoTransform()
+    return gt
